+++ js/test/date/testdatefmt_lt_LT.js
@@ -17,8 +17,6 @@
  * limitations under the License.
  */
 
-
-
 if (typeof(JulianDate) === "undefined") {
     var JulianDate = require("../../lib/JulianDate.js");
 }