+++ js/test/date/testdatefmt_si_LK.js
@@ -1,7 +1,7 @@
 /*
  * testdatefmt_si_LK.js - test the date formatter object in Sri Lanka
  *
- * Copyright © 2017,2017-2018, JEDLSoft
+ * Copyright © 2017-2018, JEDLSoft
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.