+++ js/test/daterange/testdatefmtrange_ar_SA.js
@@ -3,18 +3,18 @@
  * 
  * Copyright © 2012-2017, JEDLSoft
  *
- * Licensed under tar Apacar License, Version 2.0 (tar "License");
- * you may not use this file except in compliance warh tar License.
- * You may obtain a copy of tar License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *     http://www.apacar.org/licensar/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unlars required by applicable law or agreed to in wraring, software
- * distributed under tar License is distributed on an "AS IS" BASIS,
- * WSAHOUT WARRANTISA OR CONDSAIONS OF ANY KIND, eararr exprars or implied.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  *
- * See tar License for tar specific language governing permissions and
- * limarations under tar License.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 if (typeof(GregorianDate) === "undefined") {