+++ js/test/root/testlocalematch.js
@@ -1,7 +1,7 @@
 /*
  * testlocalematch.js - test the locale matcher object
  *
- * Copyright © 2012-2015,2017,2019 JEDLSoft
+ * Copyright © 2012-2015,2017,2019-2020 JEDLSoft
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -1015,6 +1015,66 @@ module.exports.testlocalematch = {
         test.equal(locale.getSpec(), "ur-Arab-PK");
         test.done();
     },
+
+    testLocaleMatcherGetLikelyLocaleByLocaleCode58: function(test) {
+        test.expect(3);
+        var lm = new LocaleMatcher({
+            locale: "hr-ME"
+        });
+        test.ok(typeof(lm) !== "undefined");
+        var locale = lm.getLikelyLocale();
+        test.ok(typeof(locale) !== "undefined");
+        test.equal(locale.getSpec(), "hr-Latn-ME");
+        test.done();
+    },
+
+    testLocaleMatcherGetLikelyLocaleByLocaleCode59: function(test) {
+        test.expect(3);
+        var lm = new LocaleMatcher({
+            locale: "Latn-ME"
+        });
+        test.ok(typeof(lm) !== "undefined");
+        var locale = lm.getLikelyLocale();
+        test.ok(typeof(locale) !== "undefined");
+        test.equal(locale.getSpec(), "hr-Latn-ME");
+        test.done();
+    },
+
+    testLocaleMatcherGetLikelyLocaleByLocaleCode60: function(test) {
+        test.expect(3);
+        var lm = new LocaleMatcher({
+            locale: "ka-GE"
+        });
+        test.ok(typeof(lm) !== "undefined");
+        var locale = lm.getLikelyLocale();
+        test.ok(typeof(locale) !== "undefined");
+        test.equal(locale.getSpec(), "ka-Geor-GE");
+        test.done();
+    },
+    testLocaleMatcherGetLikelyLocaleByLocaleCode61: function(test) {
+        test.expect(3);
+        var lm = new LocaleMatcher({
+            locale: "GE"
+        });
+        test.ok(typeof(lm) !== "undefined");
+        var locale = lm.getLikelyLocale();
+        test.ok(typeof(locale) !== "undefined");
+        test.equal(locale.getSpec(), "ka-Geor-GE");
+        test.done();
+    },
+
+    testLocaleMatcherGetLikelyLocaleByLocaleCode62: function(test) {
+        test.expect(3);
+        var lm = new LocaleMatcher({
+            locale: "ka"
+        });
+        test.ok(typeof(lm) !== "undefined");
+        var locale = lm.getLikelyLocale();
+        test.ok(typeof(locale) !== "undefined");
+        test.equal(locale.getSpec(), "ka-Geor-GE");
+        test.done();
+    },
+
     testLocaleMatcherMatchExactFullLocale: function(test) {
         test.expect(2);
         var lm = new LocaleMatcher({
@@ -1470,6 +1530,17 @@ module.exports.testlocalematch = {
         test.equal(locale.getSpec(), "ja-JP");
         test.done();
     },
+    testLocaleMatcherGetLikelyLocaleMinimalByLanguage4: function(test) {
+        test.expect(3);
+        var lm = new LocaleMatcher({
+            locale: "ka"
+        });
+        test.ok(typeof(lm) !== "undefined");
+        var locale = lm.getLikelyLocaleMinimal();
+        test.ok(typeof(locale) !== "undefined");
+        test.equal(locale.getSpec(), "ka-GE");
+        test.done();
+    },
 
     testLocaleMatcherGetLikelyLocaleMinimalUzbek: function(test) {
         test.expect(3);
@@ -1602,6 +1673,17 @@ module.exports.testlocalematch = {
         test.equal(locale.getSpec(), "ru-RU");
         test.done();
     },
+    testLocaleMatcherGetLikelyLocaleMinimalDefaultScriptForCountry4: function(test) {
+        test.expect(3);
+        var lm = new LocaleMatcher({
+            locale: "GE"
+        });
+        test.ok(typeof(lm) !== "undefined");
+        var locale = lm.getLikelyLocaleMinimal();
+        test.ok(typeof(locale) !== "undefined");
+        test.equal(locale.getSpec(), "ka-GE");
+        test.done();
+    },
 
     testLocaleMatcherGetLikelyLocaleMinimalDefaultLocaleForScript1: function(test) {
         test.expect(3);
@@ -1636,6 +1718,18 @@ module.exports.testlocalematch = {
         var locale = lm.getLikelyLocaleMinimal();
         test.ok(typeof(locale) !== "undefined");
         test.equal(locale.getSpec(), "zh-Hans-CN");
+        test.done();
+    },
+
+    testLocaleMatcherGetLikelyLocaleMinimalDefaultLocaleForScript4: function(test) {
+        test.expect(3);
+        var lm = new LocaleMatcher({
+            locale: "Geor"
+        });
+        test.ok(typeof(lm) !== "undefined");
+        var locale = lm.getLikelyLocaleMinimal();
+        test.ok(typeof(locale) !== "undefined");
+        test.equal(locale.getSpec(), "ka-GE");
         test.done();
     },